Crypt class
